refactor: move message banner interface to common file
