+++ frontend/src/interfaces/messageBanner.ts
@@ -1,0 +1,14 @@
+export type BannerVariant = 'warning' | 'info' | 'error' | 'success';
+
+export interface IMessageBanner {
+    message: string;
+    enabled?: boolean;
+    variant?: BannerVariant;
+    sticky?: boolean;
+    icon?: string;
+    link?: string;
+    linkText?: string;
+    plausibleEvent?: string;
+    dialogTitle?: string;
+    dialog?: string;
+}
